4 tables

+++ src/main/java/com/example/demo/repository/ExcelDataRepository.java
@@ -3,6 +3,7 @@ package com.example.demo.repository;
 import com.example.demo.dto.ExtractedDataDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExcelDataRepository extends JpaRepository<ExtractedDataDTO, Long> {
+public interface ExcelDataRepository extends JpaRepository<ExtractedDataDTO, Integer> {
 }
+
